pyproject.toml
[build-system]
requires = ["uv_build>=0.7.21,<0.8.0"]
build-backend = "uv_build"

[project]
name = "rest-filters"
version = "0.6.1"
description = "A filter backend for Django REST framework."
readme = "README.md"
authors = [{ name = "şuayip üzülmez", email = "suayip.541@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.2",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
]

[project.urls]
"Homepage" = "https://github.com/realsuayip/rest-filters"
"Bug Tracker" = "https://github.com/realsuayip/rest-filters/issues"

[dependency-groups]
dev = [
    "mypy",
    "django-stubs",
    "djangorestframework-stubs",
    "drf-spectacular",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "tox",
    "tox-uv",
]
docs = [
    "docstrfmt",
    "furo",
    "sphinx",
]

[tool.uv]
required-version = ">=0.8.6"

[tool.mypy]
strict = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
testpaths = ["tests"]
pythonpath = ". rest_filters"
addopts = """
    -Wall
    --no-migrations
    --tb=short
 """

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING",
    "raise NotImplementedError",
    "return NotImplemented",
]

[tool.coverage.html]
show_contexts = true

[tool.tox]
legacy_tox_ini = """
[tox]
envlist =
    py{310,311,312,313,314}-django{42,52}-drf{314,315,316}

[testenv]
deps =
    django42: django>=4.2,<5
    django52: django>=5.2,<6
    drf314: djangorestframework>=3.14,<3.15
    drf315: djangorestframework>=3.15,<3.16
    drf316: djangorestframework>=3.16
    django-stubs
    djangorestframework-stubs
    pytest
    pytest-cov
    pytest-django
    drf-spectacular
commands =
    pytest
"""

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    "RUF", # ruff-specific
    "C4", # flake8-comprehensions
    "C90", # mccabe
    "I", # isort
    "N", # pep8-naming
    "BLE", # flake8-blind-except
    "DTZ", # flake8-datetimez
    "DJ", # flake8-django
    "FA", # flake8-future-annotations
    "G", # flake8-logging-format
    "T20", # flake8-print
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PL", # pylint
    "PERF", # perflint
    "PIE", # flake8-pie
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "TC", # flake8-type-checking
    "UP", # pyupgrade
    "FURB", # refurb
]
fixable = ["I"]
ignore = [
    "PLR0913",
    "UP031",
    "PLW1641",
]

[tool.ruff.lint.isort]
combine-as-imports = true
section-order = [
    "future",
    "standard-library",
    "django",
    "rest_framework",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
django = ["django"]
rest_framework = ["rest_framework"]

[tool.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.extend-per-file-ignores]
"tests/*" = ["B904", "RUF012", "PLR2004"]

[tool.ruff.lint.flake8-pytest-style]
parametrize-values-type = "tuple"
parametrize-values-row-type = "tuple"
parametrize-names-type = "csv"
